Include deep linkage as a principle.
Hi there! I think this is an awesome idea, and I'll definitely be sharing this around. One thing that for me is very important when building apps is the ability to share a URL - either with myself or with others - easily. By leveraging this built in feature of the web, it makes it much easier to share, bookmark, and be a good web citizen.

Organizationally I don't know if this makes sense in the `Web applications should work everywhere` section, but just wanted to start a conversation if you're open to it.

Thanks for making this! It's much appreciated.

Dean